moved and renamed the List* types
per PEP008
